Attempting to reproduce the bug (see below) in a new project, but cannot reproduce. Interestingly, the "Project Type GUID" (in the *.sln file) is different from the MechanicsUI project. I expected them to be the same because they're both WPF apps.
Bug description: The by-distance names sometimes remain visible when the mouse leaves the canvas. This seems to happen when the mouse was hovering over the by-distance names before it left, especially if (1) the mouse moves slowly to the left and the cursor changes to the window resizer, or (2) the mouse moves quickly upwards onto the controls/info panel.