This APPEARS to fix a long-standing bug. However, I wonder if it just makes the bug a much more infrequent; perhaps it still happens when the list changes exactly while the mouse is leaving the canvas.
Bug description: The by-distance names sometimes remain visible when the mouse leaves the canvas. This seems to happen when the mouse was hovering over the by-distance names before it left, especially if (1) the mouse moves slowly to the left and the cursor changes to the window resizer, or (2) the mouse moves quickly upwards onto the controls/info panel.